chore(client/pages/Profile): split file into sub pages
