+++ src/client/src/pages/Profile/types.ts
@@ -1,0 +1,1 @@
+export type ProfilePageNames = "index" | "settings";
